@@ -1,25 +1,24 @@
 def versions = [
-    kotlinVersion                : '1.2.21',
-    supportLibVersion            : '27.1.1',
+    kotlinVersion                : '1.2.51',
+    androixVersion               : '1.0.0-beta01',
     roboelectricVersion          : '3.5',
 ]
 
 ext.versions = versions
 ext.androidConfig = [
-    agpVersion       : '3.1.3',
-    compileSdkVersion: 27,
-    buildToolsVersion: '27.0.3',
+    agpVersion       : '3.2.0-beta04',
+    compileSdkVersion: 28,
+    buildToolsVersion: '28.0.1',
     minSdkVersion    : 16,
-    targetSdkVersion : 27
+    targetSdkVersion : 28
 ]
 
 ext.deps = [
     autoCommon               : "com.google.auto:auto-common:0.8",
     androidPlugin            : "com.android.tools.build:gradle:$androidConfig.agpVersion",
-    appCompat                : "com.android.support:appcompat-v7:$versions.supportLibVersion",
-    supportCoreUtils         : "com.android.support:support-core-utils:$versions.supportLibVersion",
+    appCompat                : "androidx.appcompat:appcompat:$versions.androixVersion",
+    localBroadcastManager    : "androidx.localbroadcastmanager:localbroadcastmanager:$versions.androixVersion",
     nexusPlugin              : 'com.bmuschko:gradle-nexus-plugin:2.3.1',
-    appcompat                : "com.android.support:appcompat-v7:$versions.supportLibVersion",
     kotlinStdLib             : "org.jetbrains.kotlin:kotlin-stdlib:$versions.kotlinVersion",
     kotlinGradlePlugin       : "org.jetbrains.kotlin:kotlin-gradle-plugin:$versions.kotlinVersion",
     javaPoet                 : 'com.squareup:javapoet:1.9.0',
